requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
Django==5.1.4
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-tempus-dominus==5.1.2.17
django-timezone-field==7.0
kombu==5.6.2
pillow==11.0.0
prompt_toolkit==3.0.48
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
django-phonenumber-field==8.4.0
phonenumbers==9.0.21
django-recaptcha==4.0.0
gunicorn==23.0.0
supervisor==4.3.0
dotenv==0.9.9
psycopg2-binary==2.9.11
celery==5.6.1
